# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "Guardian"
  description: "The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation."
  version: "2.9.0"
  contact:
    name: "API developer"
    url: "https://envisionblockchain.com"
    email: "info@envisionblockchain.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
- url: /api/v1
  description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
...
    Account:
      type: object
      required:
        - username
        - role
      properties:
        username:
          type: string
        role:
          type: string
        did:
          type: string
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    HederaAccount:
      type: object
      required:
        - id
        - key
      properties:
        id:
          type: string
        key:
          type: string
...
    LogFilters:
      type: object
      properties:
        type:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        attributes:
          type: array
          items:
            type: string
        message:
          type: string
        pageSize:
          type: number
        pageIndex:
          type: number
        sortDirection:
          type: string
          enum: [ASC, DESC]
    Log:
      type: object
      properties:
        type:
          type: string
        datetime:
          type: string
        message:
          type: string
        attributes:
          type: array
          items:
            type: string
    Task:
      type: object
      properties:
        taskId:
          type: string
        expectation:
          type: number
...
paths:
...
  /demo/registeredUsers:
    get:
      tags:
        - demo
      deprecated: true
      description: Returns all registered user accounts.
      responses:
        200:
          description: All registered user accounts.
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Account'
                    - type: object
                      properties:
                        policyRoles:
                          type: object
                          properties:
                            policy:
                              type: object
                              properties:
                                id:
                                  type: string
                                name:
                                  type: string
                                version:
                                  type: string
                            userRoles:
                              type: object
                              properties:
                                did:
                                  type: string
                                role:
                                  type: string
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /demo/randomKey:
    get:
      tags:
        - demo
      deprecated: true
      description: Generates a new Hedera account with a random private key.
      security:
      - bearerAuth: []
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HederaAccount'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /demo/push/randomKey:
    get:
      tags:
        - demo
      description: Generates a new Hedera account with a random private key.
      security:
        - bearerAuth: []
      summary: Generates a new Hedera account with a random private key.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ipfs/file/{cid}:
    get:
      tags:
        - ipfs
      description: Get file from ipfs.
      summary: Get file from ipfs.
      parameters:
        - in: path
          name: cid
          schema:
            type: string
          required: true
          description: File CID.
      security:
      - bearerAuth: []
      responses:
        201:
          description: Successful operation.
          content:
            binary/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ipfs/file/:
    post:
      tags:
        - ipfs
      description: Add file to ipfs.
      summary: Add file to ipfs.
      requestBody:
        description: Data array of file.
        required: true
        content:
          binary/octet-stream:
              schema:
                type: string
                format: binary
      security:
      - bearerAuth: []
      responses:
        201:
          description: Successful operation.
          content:
            application/json:
              schema:
                description: CID of added file.
                type: string
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /logs:
    post:
      tags:
        - logs
      description: Returns logs. For users with the Standard Registry role only.
      security:
      - bearerAuth: []
      requestBody:
        description: Log filters.
        content:
          application/json:
              schema:
                $ref: '#/components/schemas/LogFilters'
      summary: Returns logs.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCount:
                    type: number
                  logs:
                    $ref: '#/components/schemas/Log'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /logs/attributes:
    get:
      tags:
        - logs
      description: Returns logs attributes. For users with the Standard Registry role only.
      parameters:
        - in: query
          name: name
          schema:
            type: string
          description: Part of name.
        - in: query
          name: existingAttributes
          schema:
            type: array
            items:
              type: string
          description: Attributes to exclude.
      security:
      - bearerAuth: []
      summary: Returns logs attributes.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...